requirements.txt
# Core dependencies for agentic AI examplesautogen-agentchat>=0.2.0autogen-core>=0.4.0autogen-ext>=0.4.0anthropic>=0.40.0python-dotenv>=1.0.0# Optional dependencies for specific examples# mcp-atlassian  # For JIRA integration# playwright  # For web automation